# improved performance when exec very long line